Fixed save file error

+++ include/GameMechanics/GameplayState.cpp
@@ -270,7 +270,7 @@ void GameplayState::update(double dt)
             glm::ivec2 enemyMZPos = data.enemyPos[i] / charSize;
 
             data.enemyPath[i] = (data.difficulty == 0) ?
-                PathFinding<BFS_Heuristic>::run({ enemyMZPos.x, enemyMZPos.y }, { mazePos.x, mazePos.y }, { data.mazeSize.x, data.mazeSize.y }, mazeEncode) :
+                PathFinding<BFS_Heuristic>::run({ enemyMZPos.x, enemyMZPos.y }, { mazePos.x, mazePos.y }, { (int)data.mazeSize.x, (int)data.mazeSize.y }, mazeEncode) :
                 PathFinding<Dijkstra>::run({enemyMZPos.x, enemyMZPos.y}, {mazePos.x, mazePos.y}, { data.mazeSize.x, data.mazeSize.y }, mazeEncode);
             data.enemyMaxStep[i] = (std::min)((unsigned long long)2, data.enemyPath[i].size() - 1);
             data.enemyStep[i] = 0;
